+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [@mjkuranda](https://github.com/mjkuranda): GameAnimation as a game resource.
 - [@mjkuranda](https://github.com/mjkuranda): Spawning an explosion animation after destroying asteroid.
 
+### Changed
+- [@mjkuranda](https://github.com/mjkuranda): RetroRenderer refactored rendering entities.
+
 ## [0.13.0] - 2023-07-26
 ### Added
 - [@mjkuranda](https://github.com/mjkuranda): Each entity has an image reference.